% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer.R
\name{tmpl_beamer}
\alias{tmpl_beamer}
\title{Beamer presentation template}
\usage{
tmpl_beamer(dir_output = getwd(), rmd_name = "beamer", open_rmd = TRUE)
}
\arguments{
\item{dir_output}{The path of the directory in which the R Markdown document
and its support files should be created.}

\item{rmd_name}{The name of the R Markdown file (without file extension).}

\item{open_rmd}{Logical, whether to open the R Markdown file after it has
been created.}
}
\description{
Beamer presentation template
}
